pyproject.toml
[project]
name = "zerotel"
version = "0.1.0"
description = "Zero-config observability SDK for Python services — auto-instruments FastAPI with traces, metrics, and logs via OpenTelemetry"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Kamalesh Somasundaram", email = "kamalesh.s7316@gmail.com" }
]
keywords = ["observability", "opentelemetry", "fastapi", "tracing", "metrics", "logging", "otel"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
    "Framework :: FastAPI",
]
requires-python = ">=3.10"

dependencies = [
    # OpenTelemetry core API + SDK
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    # OTLP exporter — sends traces/metrics to any OTEL collector
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    # Auto-instrumentation for FastAPI and outbound HTTP
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-httpx>=0.41b0",
    # Prometheus metrics exposition
    "prometheus-client>=0.17.0",
    # Structured JSON logging
    "structlog>=23.1.0",
    # Core web framework (FastAPI-first)
    "fastapi>=0.100.0",
    "starlette>=0.27.0",
]

[project.optional-dependencies]
# Async SQLAlchemy query tracing (child spans per DB call)
sqlalchemy = [
    "opentelemetry-instrumentation-sqlalchemy>=0.41b0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.28.0",
]
# Flask adapter (WSGI)
flask = [
    "opentelemetry-instrumentation-flask>=0.41b0",
    "flask>=2.3.0",
]
# Install everything
all = [
    "zerotel[sqlalchemy]",
    "zerotel[flask]",
]
# Developer / test toolchain
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.24.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/Kamalesh-Kavin/zerotel"
Repository = "https://github.com/Kamalesh-Kavin/zerotel"
Issues = "https://github.com/Kamalesh-Kavin/zerotel/issues"
Changelog = "https://github.com/Kamalesh-Kavin/zerotel/blob/main/CHANGELOG.md"
Documentation = "https://github.com/Kamalesh-Kavin/zerotel/tree/main/docs"

[build-system]
requires = ["uv_build>=0.10.10,<0.11.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "RUF"]
ignore = ["B008"]  # FastAPI Depends() in default args is intentional

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src/zerotel --cov-report=term-missing -q"

[tool.coverage.run]
source = ["src/zerotel"]
omit = ["tests/*"]
